[docs] AppConfigSchemaTable refactor and appearance fixes
